Issue 207 (#8)
This PR deals with the fact that Mikado pick was not leveraging correctly the multiple processors. This was due to the fact that the main process was taking up the job of checking transcripts and creating loci - expensive operations that acted as bottlenecks. Now the main process will only collate transcripts as GTF rows, do a minimal check on the fact that they do not have introns longer that the maximum size, and then and only then dispatch them.
Moreover, the trigger to user reduction methods in a locus has been lowered (from 10,000 to 5,000) and the first method (removal of redundant, completely contained intron chains) will be triggered before loading transcript data from the database.