WiP - Expanding uses of attributes for scoring
This is testing for use of attributes as cds_requirements, curiously, the 'cdna_length' attribute is 'swallowed' by mikado if set directly on the object, but 'something' is not. This needs to be checked as collisions on names that are handled specially should be validated at the configuration parsing stage.